+++ Source/Scene/OrthographicOffCenterFrustum.js
@@ -6,9 +6,9 @@ define([
         OrthographicOffCenterFrustum) {
     'use strict';
 
-    function DeprecatedOrthographicOffCenterFrustum() {
+    function DeprecatedOrthographicOffCenterFrustum(options) {
         deprecationWarning('OrthographicOffCenterFrustum', 'Scene/OrthographicOffCenterFrustum is deprecated. It has moved to Core/OrthographicOffCenterFrustum in 1.36. Scene/OrthographicOffCenterFrustum will be removed in 1.38.');
-        return new OrthographicOffCenterFrustum();
+        return new OrthographicOffCenterFrustum(options);
     }
 
     return DeprecatedOrthographicOffCenterFrustum;